Add: - new controller (AddItem) - 2 buttons (Done and Cancel)
AddItem has one cell
The "Done" button does not work when there is no text in the text field.
You don’t have to tap into the text field in order to bring up the keyboard, the keyboard automatically appear once the screen opens.
The "Done" and "Cancel" buttons are return you to the Checklist screen